backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python-headless==4.9.0.80
python-multipart==0.0.6
supabase==2.10.0
python-dotenv==1.0.0
numpy==1.26.3
websockets>=15.0.0

